Connected API!! connected activity, tracker, auth api

+++ Android/Wheelchair/app/src/main/java/com/softwarica/wheelchairapp/network/model/Coordinates.kt
@@ -1,5 +1,5 @@
 package com.softwarica.wheelchairapp.network.model
 
 data class Coordinates(
-    val coordinates: Array<Int>
+    val coordinates: Array<Double>
 )
